OPENEUROPA-2252: Update route and prepare for request updates.
